chore: move memory jd client to internal package
This is a refactor to move the memory jd client to an internal package
to remove usage of the memory package.
