+++ deployment/keystone/changeset/test/env_setup.go
@@ -28,8 +28,8 @@ import (
 	commontypes "github.com/smartcontractkit/chainlink/deployment/common/types"
 	"github.com/smartcontractkit/chainlink/deployment/cre/contracts"
 	"github.com/smartcontractkit/chainlink/deployment/cre/ocr3"
-	"github.com/smartcontractkit/chainlink/deployment/environment/memory"
 	envtest "github.com/smartcontractkit/chainlink/deployment/environment/test"
+	"github.com/smartcontractkit/chainlink/deployment/internal/jdtestutils"
 	"github.com/smartcontractkit/chainlink/deployment/keystone/changeset"
 	"github.com/smartcontractkit/chainlink/deployment/keystone/changeset/internal"
 	"github.com/smartcontractkit/chainlink/deployment/utils/nodetestutils"
@@ -519,10 +519,22 @@ func setupMemoryNodeTest(
 	dons.Put(newMemoryDon(c.AssetDonConfig.Name, assetNodes))
 	dons.Put(newMemoryDon(c.WriterDonConfig.Name, cwNodes))
 
-	env := memory.NewMemoryEnvironmentFromChainsNodes(
-		t.Context, lggr, blockchains, dons.AllNodes(),
+	var nodeIDs []string
+	for id := range dons.AllNodes() {
+		nodeIDs = append(nodeIDs, id)
+	}
+
+	env, err := environment.New(t.Context(),
+		environment.WithLogger(lggr),
+		environment.WithNodeIDs(nodeIDs),
+		environment.WithOffchainClient(jdtestutils.NewMemoryJobClient(dons.AllNodes())),
 	)
-	return dons, env
+	require.NoError(t, err)
+
+	// Override the blockchains with the ones we want to use
+	env.BlockChains = blockchains
+
+	return dons, *env
 }
 
 func registryChain(t *testing.T, chains map[uint64]cldf_evm.Chain) uint64 {
